Feature/project view

-added file to .gitignore
-infinite scroll for activity feed for first three loads then manual click required
-initial view for project including code to modify project and add comments and comment replies
-fixed registration bugs: multiple
-fixed image upload on profile creation
-separated features section into template for reuse
-added select2 combobox support (multi select)
-upgraded parallax version from 1.7.3 to 1.10.2
-bug fixes in home view
-several fixes to unit and integration tests
-Added cache invalidation for most popular projects when projects are updated
-activity times in activity feed are relative
-added several custom tags to tag lib: booleanOut, userOwnsProject, userCanMakeChangesToProject, projectVisibility, dateDiff and getUserProfilePic with ability to ask for multiple sizes of the profile pic. Sizes supported are: xs, small, medium, large, x-large, xx-large and xxx-large. You can also request whether the image has a border (sticker class)
-added several images
-added code that increments view counts on projects

See merge request !7

+++ grails-app/jobs/edu/wustl/cielo/EmailSenderJob.groovy
@@ -1,6 +1,7 @@
 package edu.wustl.cielo
 
 import groovy.util.logging.Slf4j
+import com.mailjet.client.MailjetResponse
 
 @Slf4j
 class EmailSenderJob {
@@ -25,18 +26,18 @@ class EmailSenderJob {
 
         //Registration emails to send
         RegistrationEmail.findAllByAttemptsLessThan(3).each { regEmail ->
-            emailService.sendEmail(regEmail.toAddresses,
+            MailjetResponse response = emailService.sendEmail(regEmail.toAddresses,
                         regEmail.subject, regEmail.plainMessage, regEmail.htmlMessage)
 
-            registrationService.deleteRegistrationEmail(regEmail)
+            if (response?.getStatus() == 200) registrationService.deleteRegistrationEmail(regEmail)
         }
 
         //contact us emails to send
         ContactUsEmail.findAllByAttemptsLessThan(3).each { contactUsEmail ->
-            emailService.sendEmail(contactUsEmail.toAddresses,
+            MailjetResponse response = emailService.sendEmail(contactUsEmail.toAddresses,
                     contactUsEmail.subject, contactUsEmail.plainMessage, contactUsEmail.htmlMessage)
 
-            communicationsService.deleteContactUsEmail(contactUsEmail)
+            if (response?.getStatus() == 200) communicationsService.deleteContactUsEmail(contactUsEmail)
         }
     }
 }
